+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - `XIT SET`: Make the sidebar button list reorderable
 - `XIT SORT`: Make the sorting mode list reorderable
 - `inv-compress-inventory-info`: Add a small right padding to the unload button
+- `inv-shorten-storage-types`: Use the short type labels from the base game instead of custom ones
 - Ignore planetary infrastructure inventories in all features
 
 ### Fixed